backend/requirements.txt
absl-py==2.3.1
aiofiles==25.1.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
audioread==3.0.1
bidict==0.23.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
decorator==5.2.1
exceptiongroup==1.3.0
fastapi==0.119.1
flatbuffers==25.9.23
fonttools==4.60.1
h11==0.16.0
idna==3.11
importlib_metadata==8.7.0
importlib_resources==6.5.2
joblib==1.5.2
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.43.0
matplotlib==3.9.4
mediapipe==0.10.14
ml_dtypes==0.5.3
msgpack==1.1.2
numba==0.60.0
numpy==1.24.3
opencv-contrib-python==4.9.0.80
opencv-python==4.9.0.80
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
protobuf==4.25.8
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-engineio==4.12.3
python-multipart==0.0.20
python-socketio==5.14.2
requests==2.32.5
scikit-learn==1.6.1
scipy==1.8.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
soundfile==0.13.1
soxr==1.0.0
starlette==0.48.0
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
websockets==15.0.1
wsproto==1.2.0
zipp==3.23.0
py-feat==0.5.1
